% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_kARTe.R
\name{create_kARTe}
\alias{create_kARTe}
\title{Get spatial data from metrics.patches output}
\usage{
create_kARTe(
  location = NULL,
  bbox = NULL,
  edge_buffer = 0,
  river_linewidth = 10,
  roads_linewidth = 2,
  color_palette = "original",
  color_water = "black",
  label = "",
  gfont_name = "Lexend",
  gfont_width = NULL,
  gfont_size = 60,
  show_plot = TRUE,
  export = TRUE,
  output_dir = "",
  filename = "map",
  timestamp = TRUE,
  file_format = c("svg", "png"),
  file_width = 80,
  file_height = 80,
  file_units = "cm",
  file_dpi = 300
)
}
\arguments{
\item{location}{charatcer string, define a location for automatic bbox estimation.
If NULL (default) the provided bbox will be used instead}

\item{bbox}{numeric vector(xmin, xmax, ymin, ymax), define a bbox instead of a location string.
Only used when location == NULLs}

\item{edge_buffer}{numeric vector(x, y), define a buffer for x and y direction.
Map will be clipped by defined buffers before plotting.}

\item{river_linewidth}{numeric value, river linewidth in plot}

\item{roads_linewidth}{numeric value, roads linewidth in plot}

\item{color_palette}{charatcer, name of the color palette.
Choices are currently: "original", "blend", "synth".
Optionally, a vector with colors can be passed}

\item{color_water}{character, single color for water}

\item{label}{character, label to put below the plotting area}

\item{gfont_name}{character, name of a google font to use for the label}

\item{gfont_width}{numeric, optional width specification for google font}

\item{gfont_size}{numeric, label font size in plot}

\item{show_plot}{boolean, if TRUE (default) plot is shown in the RStudio Plots pane}

\item{export}{boolean, if TRUE (default) plot will be written to the define \code{file_format}}

\item{output_dir}{character, output directory for plot export}

\item{filename}{character, file name for plot export}

\item{timestamp}{boolean, if TRUE (default) a timestamp will be pasted to the defined \code{filename}}

\item{file_format}{character vector, define one or multiple file formats (e.g. c("svg", "png"))}

\item{file_width}{numeric, width of exported file}

\item{file_height}{numeric, height of exported file}

\item{file_units}{character, units specification for \code{file_width} and \code{file_height} , defaults to "cm"}

\item{file_dpi}{numeric, DPI of exported file}
}
\value{
ggplot object
}
\description{
Create kARTe
}
\details{
Create a kARTe plot and files
}
\examples{
\dontrun{

# Create a kARTe with bbox:
xmin=9.871
xmax=9.971
ymin=51.51
ymax=51.57

bbox = c(xmin, xmax, ymin, ymax)
edge_buffer = c(0.01, 0.005)

create_kARTe(bbox = bbox,
             edge_buffer = edge_buffer,
             label = "GÖTTINGEN",
             gfont_width = 500,
             output_dir = "",
             filename = "map")

 # Create a kARTe with location name
create_kARTe(location = "Göttingen",
             edge_buffer = edge_buffer,
             label = "GÖTTINGEN",
             gfont_width = 500,
             output_dir = "",
             filename = "map")
}

}
