% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_box.R
\name{get_box}
\alias{get_box}
\title{get_box}
\usage{
get_box(location, bbox)
}
\arguments{
\item{location}{character}

\item{bbox}{numeric vector (xmin, xmax, ymin, ymax)}
}
\value{
bbox object containg bounding box coordinates in the order ymin, xmin, ymax, xmax
}
\description{
get_box
}
