% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_feature.R
\name{get_osm_feature}
\alias{get_osm_feature}
\title{get_osm_feature}
\usage{
get_osm_feature(box, key, value)
}
\arguments{
\item{box}{bounding box object}

\item{key}{character}

\item{value}{character}
}
\description{
get_osm_feature
}
