% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_roads_to_tiles.R
\name{morph_roads_to_tiles}
\alias{morph_roads_to_tiles}
\title{morph_roads_to_tiles}
\usage{
morph_roads_to_tiles(roads)
}
\arguments{
\item{roads}{roads layer}
}
\description{
morph_roads_to_tiles
}
